feat: Split milestone frequency seconds counter to a custom react hook
